# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Absensi Mahasiswa Geolocation API",
        "description": "API untuk sistem absensi mahasiswa berbasis geolocation. Mahasiswa dapat melakukan absensi dengan validasi lokasi GPS.",
        "contact": {
            "name": "Admin Support",
            "email": "admin@kampus.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server v1"
        }
    ],
    "paths": {
        "/attendance": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Submit absensi",
                "description": "Mahasiswa submit absensi dengan koordinat GPS. Status akan 'hadir' jika dalam radius lokasi, 'ditolak' jika di luar radius.",
                "operationId": "storeAttendance",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "description": "ID jadwal yang akan diabsen",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "latitude": {
                                        "description": "Latitude posisi user",
                                        "type": "number",
                                        "format": "float",
                                        "example": -6.2
                                    },
                                    "longitude": {
                                        "description": "Longitude posisi user",
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.816666
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Absensi berhasil diproses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "hadir",
                                                "ditolak"
                                            ],
                                            "example": "hadir"
                                        },
                                        "distance": {
                                            "description": "Jarak dari lokasi dalam meter",
                                            "type": "number",
                                            "format": "float",
                                            "example": 45.23
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi berhasil dicatat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Absensi gagal (di luar waktu jadwal atau sudah absen)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "distance": {
                                            "type": "number",
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi hanya dapat dilakukan pada waktu jadwal aktif"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attendance/history": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Riwayat absensi user",
                "description": "Mendapatkan riwayat absensi user yang sedang login (dengan pagination)",
                "operationId": "attendanceHistory",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan riwayat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "schedule": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "course_name": {
                                                                "type": "string",
                                                                "example": "Pemrograman Web"
                                                            },
                                                            "start_time": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "end_time": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "hadir"
                                                    },
                                                    "distance": {
                                                        "type": "number",
                                                        "example": 45.23
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 75
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/schedules/today": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Jadwal hari ini",
                "description": "Mendapatkan daftar jadwal untuk hari ini",
                "operationId": "todaySchedules",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan jadwal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "course_name": {
                                                        "type": "string",
                                                        "example": "Pemrograman Web"
                                                    },
                                                    "location_name": {
                                                        "type": "string",
                                                        "example": "Gedung A - Fakultas Teknik"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token yang didapat dari endpoint login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}